MB-52931 Allow drop index on source during rebalance
During a rebalance, if a drop index command is received,
then indexer will attempt to drop the index on source
node if:

a. Bucket is not a part of rebalance - Then the index
will be dropped immediately

b. Bucket is undergoing rebalance - Then the index will
be dropped on source node only after the transfer of
the shard data to destination node (only transfer - not
the entire rebalance of bucket) is complete. Indexer
can not allow drop on source nodes as plasma expects the
shard metadata to be consistent during the transfer.

In such a case, the drop is processed by lifecycleMgr::janitor
component. Janitor checks the delete command tokens periodically
and then drops the tokens. To facilitate drop during rebalance,
the rebalanceRunning check at indexer is skipped based on
the allowDDLDuringRebalance flag. Also, janitor is updated to
invoke cleanup when rebalance is running if allowDDLDuringRebalance
is set to true for serverless deployments.

The index drop at destination is handled separately (as destination
node can skip creation of index if it knows that the index is
already dropped during rebalance)

+++ secondary/indexer/indexer.go
@@ -3715,20 +3715,24 @@ func (idx *indexer) handleDropIndex(msg Message) (resp Message) {
 	}
 
 	if idx.rebalanceRunning || idx.rebalanceToken != nil {
+		if idx.canAllowDDLDuringRebalance() && msg.(*MsgDropIndex).GetMsgType() == CLUST_MGR_DROP_INDEX_DDL {
+			logging.Infof("Indexer::handleDropIndex Allowing drop index during rebalance for "+
+				"index defnId: %v, instId: %v", indexInst.Defn.DefnId, indexInst.InstId)
+		} else {
+			reqCtx := msg.(*MsgDropIndex).GetRequestCtx()
+			if reqCtx.ReqSource == common.DDLRequestSourceUser {
+				errStr := fmt.Sprintf("Indexer Cannot Process Drop Index - Rebalance In Progress")
+				logging.Errorf("Indexer::handleDropIndex %v", errStr)
 
-		reqCtx := msg.(*MsgDropIndex).GetRequestCtx()
-		if reqCtx.ReqSource == common.DDLRequestSourceUser {
-			errStr := fmt.Sprintf("Indexer Cannot Process Drop Index - Rebalance In Progress")
-			logging.Errorf("Indexer::handleDropIndex %v", errStr)
-
-			if clientCh != nil {
-				clientCh <- &MsgError{
-					err: Error{code: ERROR_INDEXER_REBALANCE_IN_PROGRESS,
-						severity: FATAL,
-						cause:    errors.New(errStr),
-						category: INDEXER}}
+				if clientCh != nil {
+					clientCh <- &MsgError{
+						err: Error{code: ERROR_INDEXER_REBALANCE_IN_PROGRESS,
+							severity: FATAL,
+							cause:    errors.New(errStr),
+							category: INDEXER}}
+				}
+				return
 			}
-			return
 		}
 	}
